+++ src/abstractClasses/LaserTower.cpp
@@ -5,7 +5,7 @@ LaserTower::LaserTower(EnemyManager* enemyManager, Coords coords): Tower(enemyMa
     this->enemyManager = enemyManager;
     this->towerCoords = coords;
 
-    float exp = 0;
+    double exp = 0;
     for (int i = 0; i <= MAX_TOWER_LEVEL; i++){
         listExpForLvls.push_back(exp);
         exp += 100;
@@ -14,7 +14,7 @@ LaserTower::LaserTower(EnemyManager* enemyManager, Coords coords): Tower(enemyMa
     setAttackTimer();
 }
 
-LaserTower::LaserTower(EnemyManager* enemyManager, float damage, float radius, float attackSpeed, int level, Coords coords):
+LaserTower::LaserTower(EnemyManager* enemyManager, double damage, double radius, double attackSpeed, int level, Coords coords):
 Tower(enemyManager, damage, radius, attackSpeed, level, coords){
 
     this->enemyManager = enemyManager;
@@ -24,7 +24,7 @@ Tower(enemyManager, damage, radius, attackSpeed, level, coords){
     this->level = level;
     this->towerCoords = coords;
 
-    float exp = 0;
+    double exp = 0;
     for (int i = 0; i <= MAX_TOWER_LEVEL; i++){
         listExpForLvls.push_back(exp);
         exp += 100;
